1.捕获运行时异常并打印 2.添加清空目录功能（不清空excel和jar） 3.解决excel读取空白行的BUG.

+++ src/main/java/com/nstc/dbwriter/control/ClearTemp.java
@@ -18,7 +18,7 @@ import com.nstc.dbwriter.util.WriteUtil;
  */
 public class ClearTemp {
     
-    public static void main(String[] args) {
+    public static void clear() {
         //删除model
         File modelDir = new File(InnerSettings.TEST_MODEL_DIR);
         clearDir(modelDir);
@@ -47,6 +47,12 @@ public class ClearTemp {
             for (File file : files) {
                 if(file.isFile()) {
                     if(".gitkeep".equals(file.getName())){
+                        continue;
+                    }
+                    if(file.getName().endsWith(".xlsx")){
+                        continue;
+                    }
+                    if(file.getName().endsWith(".jar")){
                         continue;
                     }
                     file.delete();
